refactor(timeline): split items-store into normalize + indexes modules
items-store.ts grew to 1746 lines mixing the Zustand store with audio EQ
normalization, frame-field rounding, derived index builders, and subtitle
cue helpers. Extract:

- items-store-normalize.ts: normalizeFrameFields, normalizeItemUpdates,
  normalizeTrack, subtitle cue trim helpers, roundFrame helpers.
- items-store-indexes.ts: buildItemsByTrackId, buildItemById,
  buildLinkedItemsByItemId, withItemIndexes, buildRippleShiftByItemId,
  getTransitionLinkedIds, and supporting predicates.

items-store.ts (now 897 lines) contains the store factory plus its
media-dependency and in/out-point side-effect subscriptions. No behavior
or public API changes — useItemsStore remains the only export and is
imported unchanged by all 65 consumers.
